The weaponRNG method returns a random weapon to be assigned to the new private String varable found in all survivor classes The method is called within the populateArray() method and the return is passed as an argument when each survivor is created.
Accuracy and damage are calculated within the survivorAttackZombie() method.
As you can see in the weaponRNG() method, the weapon name, damage, and accuracy are all stored in the return.
The return is split into seperate data values in the attack method.
If statements were used to check RNG of weapon hitting.
setAttack() method used to modify their attack based on the random weapon they got.
    
Feel free to make any changes that you see fit.